fixed rates, added DTO

+++ src/main/java/com/example/CompConf/CompConf/Model/GraphicsCard.java
@@ -18,13 +18,13 @@ public class GraphicsCard{
     private String model;
     private double price;
     private boolean avaliability;
-    private short rates;
+    private int rates;
 
     public GraphicsCard() {
 
     }
 
-    public GraphicsCard(Long ID, String brand, String model, double price, boolean avaliability, short rates) {
+    public GraphicsCard(Long ID, String brand, String model, double price, boolean avaliability, int rates) {
         this.ID = ID;
         this.brand = brand;
         this.model = model;
@@ -73,11 +73,11 @@ public class GraphicsCard{
         this.avaliability = avaliability;
     }
 
-    public short getRates() {
+    public int getRates() {
         return rates;
     }
 
-    public void setRates(short rates) {
+    public void setRates(int rates) {
         this.rates = rates;
     }
 }
