+++ src/main/java/com/example/CompConf/CompConf/Model/Motherboard.java
@@ -14,14 +14,14 @@ public class Motherboard{
     private String model;
     private double price;
     private boolean avaliability;
-    private short rates;
+    private int rates;
     /*
          Do more specification laterrr // done in documentation
          */
     public Motherboard() {
     }
 
-    public Motherboard(Long ID, String brand, String model, double price, boolean avaliability, short rates) {
+    public Motherboard(Long ID, String brand, String model, double price, boolean avaliability, int rates) {
         this.ID = ID;
         this.brand = brand;
         this.model = model;
@@ -70,11 +70,11 @@ public class Motherboard{
         this.avaliability = avaliability;
     }
 
-    public short getRates() {
+    public int getRates() {
         return rates;
     }
 
-    public void setRates(short rates) {
+    public void setRates(int rates) {
         this.rates = rates;
     }
 }
