+++ src/main/java/com/example/CompConf/CompConf/Model/Processor.java
@@ -15,7 +15,7 @@ public class Processor{
     private String model;
     private double price;
     private boolean avaliability;
-    private short rates;
+    private int rates;
 
     /*
      Do more specification laterrr // done in documentation
@@ -25,7 +25,7 @@ public class Processor{
 
     }
 
-    public Processor(Long ID, String brand, String model, double price, boolean avaliability, short rates) {
+    public Processor(Long ID, String brand, String model, double price, boolean avaliability, int rates) {
         this.ID = ID;
         this.brand = brand;
         this.model = model;
@@ -66,11 +66,11 @@ public class Processor{
         this.avaliability = avaliability;
     }
 
-    public short getRates() {
+    public int getRates() {
         return rates;
     }
 
-    public void setRates(short rates) {
+    public void setRates(int rates) {
         this.rates = rates;
     }
